refactor: simplify state management in synth preset manager and related components
